refactor: move utility methods to new AgreementsBaseSpec file and indent tests
